mainRestEasy/requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
requests==2.28.1
RPi.GPIO==0.7.1
urllib3==1.26.13
